pyproject.toml
[project]
name = "linker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-docx",
    "Pillow",
    "langchain-anthropic",
    "langchain-google-genai",
    "python-dotenv",
    "ipywidgets",
    "openpyxl>=3.1.5",
    "docxtpl",
    "pdfplumber>=0.11.9",
    "easyocr>=1.7.2",
    "pytesseract>=0.3.13",
]
